+++ pkg/progress/progress_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/osbuild/bootc-image-builder/bib/pkg/progress"
+	"github.com/osbuild/image-builder-cli/pkg/progress"
 )
 
 func TestProgressNew(t *testing.T) {
